feat(@zorg/contracts): watcher for dev
closes #49
