+++ mprocs.local.yaml
@@ -1,10 +1,13 @@
 procs:
   contracts:
     cwd: packages/contracts
-    shell: wait-port 5050 && sozo build --typescript --bindings-output ../client/src/lib/dojo_bindings/ && sozo inspect && sozo migrate && sozo inspect && torii --world "0x059cbdb7b3b6f4b7828e8b50f2c1bfd6cab76212be4492ec5fee80671a8355b2" --http.cors_origins "*"
+    shell: wait-port 5050 && sozo build --typescript --bindings-output ../client/src/lib/dojo_bindings/ && sozo inspect && sozo migrate && sozo inspect && bun run dev:watch
   client:
     cwd: packages/client
     shell: bun run dev
+  torii:
+    cwd: packages/contracts
+    shell: wait-port 5050 && torii --world "0x059cbdb7b3b6f4b7828e8b50f2c1bfd6cab76212be4492ec5fee80671a8355b2" --http.cors_origins "*"
   katana:
     cwd: packages/contracts
     shell: katana --dev --dev.no-fee --http.cors_origins "*"
