+++ packages/contracts/manifest_slot.json
@@ -2092,7 +2092,7 @@
     },
     {
       "address": "0x7a0563afc0f01e47d2982b23435d7d8552f2bb48a2a8da2c9046dff72e6b3ff",
-      "class_hash": "0x2846d32b4124139944a3afaf9576d48303385c8c050df703087ac2f60dd96f5",
+      "class_hash": "0x9750024285ff948394cc4b77d38b8c85e76d0c125db421f4c77336365ca8bf",
       "abi": [
         {
           "type": "impl",
