+++ packages/contracts/package.json
@@ -6,15 +6,17 @@
 		"bun": "latest"
 	},
 	"scripts": {
-		"slot:deploy": "FORCE_COLOR=3 bun scripts/slot.deploy.ts",
-		"slot:upgrade": "FORCE_COLOR=3 bun scripts/slot.upgrade.ts",
-		"slot:watch": "FORCE_COLOR=3 bun scripts/slot.watch.ts"
+		"slot:deploy": "FORCE_COLOR=3 bun scripts/slot.deploy.ts --mode slot",
+		"slot:upgrade": "FORCE_COLOR=3 bun scripts/slot.upgrade.ts --mode slot",
+		"slot:watch": "FORCE_COLOR=3 bun scripts/slot.watch.ts --mode slot",
+		"dev:watch": "FORCE_COLOR=3 bun scripts/dev.watch.ts --mode dev"
 	},
 	"devDependencies": {
 		"@clack/core": "^0.4.1",
 		"@clack/prompts": "^0.10.0",
 		"ansicolor": "^2.0.3",
 		"dettle": "^1.0.5",
+		"mri": "^1.2.0",
 		"smol-toml": "^1.3.1"
 	}
 }
