+++ packages/contracts/scripts/slot.upgrade.ts
@@ -1,4 +1,5 @@
-import { deploymentComplete, runContractDeployment } from "./common";
+import { deploymentComplete } from "./common";
+import { runContractDeployment } from "./slot.common";
 
 await runContractDeployment();
 deploymentComplete();
